fix(cli,core): static subagent path drops its dead dynamic import; routeDescriptors typed + wired-through test
static-route-execution.test.ts edited ONLY to replace its local helper
definitions with imports from test/helpers/static-modules-fixture.ts —
zero assertion/test-body changes (sanctioned exception).
